feat!: rig-effect-log, record and replay as their own crate
EffectLog, LogHeader, EFFECT_LOG_FORMAT and stable_hash leave
rig_core::effect; EffectLogRecorder (a rig_bus::Recorder) and
EffectLogReplayer leave rig-bus. The new crate depends on rig-core and
rig-bus and is the persistence story over both: what a host that saves
and restores in-flight effects depends on. EffectRecord stays vocabulary
in rig-core. rig-agent's record_effects/take_effect_log/run_spec_hash/
check_replayable keep their signatures over the new paths.

The record/replay tests over a live bus move from rig-bus's tests to the
log crate (a dev-dependency back onto rig-bus builds the crate twice and
the trait impls no longer line up, so the tests live where the types
do); the helpers they use are copies. The log round-trip test moves with
EffectLog. The guards scan four crates; the dependency-graph guard
asserts rig-effect-log is runtime-free and sits above rig-bus, and
neither rig-core nor rig-bus depends on it; provider_layout exempts the
package; the wasm matrix checks it; the facade exposes rig::effect_log.
The trybuild diagnostic re-pinned (the impl list no longer names the
replayer). Publish dry-run fails on the registry's missing rig-bus, as
expected until rig-bus publishes.

Migration: rig_core::effect::{EffectLog, LogHeader, EFFECT_LOG_FORMAT,
stable_hash} and rig_bus::{EffectLogRecorder, EffectLogReplayer} →
rig_effect_log::…
