+++ crates/rig-agent/Cargo.toml
@@ -26,6 +26,7 @@ futures = { workspace = true }
 http = { workspace = true }
 indexmap = { workspace = true }
 rig-bus = { path = "../rig-bus", version = "0.42.0" }
+rig-effect-log = { path = "../rig-effect-log", version = "0.42.0" }
 rig-core = { path = "../rig-core", version = "0.42.0", default-features = false }
 rig-derive = { path = "../rig-derive", version = "0.42.0", optional = true }
 schemars = { workspace = true }
