+++ crates/rig-agent/src/agent/completion.rs
@@ -12,8 +12,9 @@ use crate::{
 };
 use rig_bus::{BusDriver, Dispatcher, ModelHandle};
 use rig_core::completion::ModelRef;
-use rig_core::effect::{EffectLog, HandlerDescriptor, Key, family};
+use rig_core::effect::{HandlerDescriptor, Key, family};
 use rig_core::id::ConversationId;
+use rig_effect_log::EffectLog;
 
 use super::bus::AgentBus;
 use rig_core::{message::ToolChoice, wasm_compat::WasmCompatSend};
@@ -515,7 +516,7 @@ impl Agent {
     /// carries in its header and what [`check_replayable`](Self::check_replayable)
     /// compares.
     pub fn run_spec_hash(&self) -> u64 {
-        rig_core::effect::stable_hash(&self.config.run_spec()).unwrap_or_default()
+        rig_effect_log::stable_hash(&self.config.run_spec()).unwrap_or_default()
     }
 
     fn stamp(&self, mut log: EffectLog) -> EffectLog {
@@ -529,7 +530,7 @@ impl Agent {
     /// recorded family. Refused up front, with both sides in the message,
     /// rather than at the record where the run would have diverged.
     pub fn check_replayable(&self, log: &EffectLog) -> Result<(), rig_core::error::ErrorReport> {
-        rig_bus::EffectLogReplayer::check_header(log)?;
+        rig_effect_log::EffectLogReplayer::check_header(log)?;
         if let Some(recorded) = log.header.run_spec {
             let mine = self.run_spec_hash();
             if recorded != mine {
