+++ crates/rig-effect-log/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "rig-effect-log"
+version.workspace = true
+edition = { workspace = true }
+license = "MIT"
+description = "Record and replay for rig's effect bus: the effect log, its header, the recorder a driver writes to and the replayer that answers from the record."
+repository = "https://github.com/0xPlaygrounds/rig"
+
+[package.metadata.docs.rs]
+all-features = true
+rustdoc-args = ["--cfg", "docsrs"]
+
+[lints]
+workspace = true
+
+[dependencies]
+rig-bus = { path = "../rig-bus", version = "0.42.0" }
+rig-core = { path = "../rig-core", version = "0.42.0", default-features = false }
+serde = { workspace = true, features = ["derive"] }
+serde_json = { workspace = true }
+
+[dev-dependencies]
+futures = { workspace = true }
+rig-core = { path = "../rig-core", version = "0.42.0", features = ["test-utils"] }
+tokio = { workspace = true, features = ["full"] }
