+++ crates/rig-effect-log/src/tests.rs
@@ -1,0 +1,65 @@
+use rig_core::{
+    completion::{AssistantContent, CompletionRequest, CompletionResponse, Message, Usage},
+    effect::{EffectId, EffectKind, EffectRecord, HandlerKey, Outcome},
+    error::{ErrorKind, ErrorReport},
+    tool::ToolContext,
+};
+
+use super::*;
+
+fn request() -> CompletionRequest {
+    CompletionRequest {
+        model: None,
+        chat_history: vec![Message::user("hi")],
+        documents: vec![],
+        tools: vec![],
+        temperature: None,
+        max_tokens: None,
+        tool_choice: None,
+        additional_params: None,
+        output_schema: None,
+        record_telemetry_content: false,
+    }
+}
+
+#[test]
+fn effect_record_and_log_round_trip() {
+    let log: EffectLog = EffectLog::from_records(vec![
+        EffectRecord {
+            id: EffectId::from_raw(1),
+            key: HandlerKey::from("model"),
+            kind: EffectKind::Completion {
+                request: request(),
+                stream: false,
+            },
+            outcome: Ok(Outcome::Completion(CompletionResponse::new(
+                vec![AssistantContent::text("hi")],
+                Usage::new(),
+                "mock",
+            ))),
+            events: None,
+        },
+        EffectRecord {
+            id: EffectId::from_raw(2),
+            key: HandlerKey::from("tool:add"),
+            kind: EffectKind::ToolCall {
+                name: "add".into(),
+                args: "{}".into(),
+                context: ToolContext::new(),
+            },
+            outcome: Err(ErrorReport::new(ErrorKind::Timeout, "slow")),
+            events: None,
+        },
+    ]);
+    let json = serde_json::to_string(&log).expect("serializes");
+    let back: EffectLog = serde_json::from_str(&json).expect("deserializes");
+    assert_eq!(
+        serde_json::to_string(&back).expect("serializes"),
+        json,
+        "log round trip"
+    );
+    assert_eq!(back.len(), 2);
+    assert_eq!(back[0].id, EffectId::from_raw(1));
+    assert_eq!(back[1].key.as_str(), "tool:add");
+    assert!(matches!(&back[1].outcome, Err(report) if report.kind == ErrorKind::Timeout));
+}
